bitcoin: move block test to tests/
Move segwit_block to tests/ because it include_bytes'
data from tests/, which we exclude when publishing.

Test used two private helpers. Work::from_hex replaces
pow::test_utils::u64_to_work, block_base_size is small
enough to duplicate.
